Migrate DynamicCSVRenderer -> dso CSVRenderer (as it has no dynamic API parts)
